feat(layout): display app icon and restructure component organization
This commit renames the Layout and LayoutHeader components, moving them to a higher directory level for better organization. The import paths have been updated accordingly. Additionally, an app icon is now displayed in the LayoutHeader for enhanced branding. These changes aim to improve code clarity and maintainability while ensuring a consistent structure across component files.
