+++ src/components/Layout/index.ts
@@ -1,1 +1,1 @@
-export * from './components/Layout'
+export * from './Layout'
